rename and move tests to better separate python api tests vs. repository/orchestration tests
