+++ shyft/tests/api/TestCalendarAndTime.py
@@ -1,4 +1,4 @@
-from shyft import api
+﻿from shyft import api
 import unittest
 import datetime as dt
