+++ shyft/tests/tests.pyproj
@@ -21,6 +21,7 @@
     <PtvsTargetsFile>$(MSBuildExtensionsPath32)\Microsoft\VisualStudio\v$(VisualStudioVersion)\Python Tools\Microsoft.PythonTools.targets</PtvsTargetsFile>
   </PropertyGroup>
   <ItemGroup>
+    <Compile Include="test_arome_respository.py" />
     <Compile Include="test_gis_region_model_repository.py">
       <SubType>Code</SubType>
     </Compile>
